backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.6.1
faiss-cpu==1.8.0
scipy==1.13.0